V1：优化退出登录，会话失效的情况直接返回退出登录成功

+++ web-api/A-api/src/main/java/net/cocotea/admin/api/controller/system/LoginController.java
@@ -33,13 +33,14 @@ public class LoginController {
         return ApiResult.ok(r);
     }
 
-    @SaCheckLogin
     @Mapping("/logout")
     public ApiResult<String> logout() {
-        // 删除权限缓存
-        redisService.delete(String.format(RedisKey.USER_PERMISSION, StpUtil.getLoginId()));
-        redisService.delete(String.format(RedisKey.ONLINE_USER, StpUtil.getLoginId()));
-        StpUtil.logout();
+        if ( StpUtil.isLogin() ) {
+            // 删除权限缓存
+            redisService.delete(String.format(RedisKey.USER_PERMISSION, StpUtil.getLoginId()));
+            redisService.delete(String.format(RedisKey.ONLINE_USER, StpUtil.getLoginId()));
+            StpUtil.logout();
+        }
         return ApiResult.ok();
     }
